backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
motor>=3.4.0
pydantic>=2.7.0
python-dotenv>=1.0.1
email-validator>=2.1.0
